Rudimentary vertical navbar (#5)
This Navbar needs better styling and also the links need to dynamically update depending on user role and URL location. Design your pages with this navbar on the side because this is what they will look like in the end. If you place your page in a proper route in App.tsx, the navbar will be added automatically, you don't have to do anything else